refactor: Study Record Page 파일 분리
